move most libusb code to its own file
